+++ web-client/src/App.js
@@ -2,14 +2,15 @@
 import { useEffect } from 'react';
 import NavBar from './Components/NavBars/welcomePageNavBar';
 import { Route, Routes } from 'react-router-dom';
-import Home from './pages/Home'
+import Welcome from './pages/Welcome'
 import SignIn from './pages/SignIn';
 import SignUp from './pages/SignUp';
-import Dashboard from './pages/dashboard';
+import Home from './pages/home';
 import { UserAuthContextProvider } from "./firebase/UserAuthContext";
 import ProtectedRoute from "./firebase/protectedRoute";
 import { RoleSelection } from './pages/roleSelection';
 import EmployerSignUp from './pages/employerSignUp'
+//import Welcome from './pages/Welcome';
 
 //import api_key from './secrets'
 
@@ -30,10 +31,10 @@ const App = () => {
 
     <UserAuthContextProvider>
       <Routes>
-        <Route path="/" element={<Home/>} />
+        <Route path="/" element={<Welcome/>} />
         <Route path="/sign-in" element={<SignIn/>} />
         <Route path="/sign-up" element={<SignUp/>} />
-        <Route path="/dashboard" element={<ProtectedRoute><Dashboard/></ProtectedRoute>}/>
+        <Route path="/home" element={<ProtectedRoute><Home/></ProtectedRoute>}/>
         <Route path="/role-selection" element={<RoleSelection/>} />
         <Route path="/employer-sign-up" element={<EmployerSignUp/>} />
       </Routes>
