+++ web-client/src/pages/SignUp.js
@@ -25,7 +25,7 @@ const Signup = () => {
     handleSave();
     try {
       await signUp(email, password);
-      navigate("/dashboard");
+      navigate("/home");
     } catch (err) {
       setError(err.message);
     }
