--- web-client/src/pages/dashboard.js
@@ -1,16 +1,0 @@
-
-import NavBar from '../Components/NavBars/authorizedNavBar';
-import {Button} from 'react-bootstrap';
-
-export default function Dashboardashboard() {
-
-
-  return (
-    <>
-    <NavBar/>
-    <div>Dashboard</div>
-
-    </>
-  )
-}
-
